requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
torch>=2.0.0
transformers>=4.30.0
soundfile>=0.12.1
piper-tts>=1.2.0
Levenshtein>=0.21.0
numpy>=1.24.0